Add solution to ASCII table program
Credit for exercise: https://cscircles.cemc.uwaterloo.ca/8-remix/

This was a very interesting problem, in that it wasn't particular
complex from a technical aspect, but it did required a decent amount of
forethought in terms of how to deal with formatting and displaying of
the ASCII information.

It was interesting to deal with the dichotomy between the visual
presentation of the ASCII information which the brain sees as a list of
ASCII characters and their corresponding number codes below them and the
reality of how this information is put together in Python. In this case,
the information is built up as a continguous string of ASCII characters
with a line break, followed by another contiguous string of ASCII code
numbers. Breaking through the logical barrier of how I was
seeing the structure of the table I wanted to produce and converting
that to a new way of structuring the data so it can be printed in the
desired format was the biggest challenge here.

The other challenge I encountered was to work through the vast majority
of the implementation of this solution on paper first in words as
opposed to even pseudo-code. This really helped me with solving the
problem at hand in earnest, instead of moving between problem solving
and implementation. There was still some massaging of the code and logic
to be done once I'd written some Python code, but that vast majority of
the logic (particularly around formatting) was complete before writing a
single line of Python code.

Another thing I thought about when writing this program came from what
I've learned from approaching novel math problems: what are the
invariables and what are the variables? In this case, I identified the
"chr: " and "asc: " strings at the beginning of each line as clear
invariables in the expected output, along with the more subtle
invariable of the width of invisible box that each ASCII character and
ASCII number code had to fit in (in this case, it was equal to the
length of the largest ASCII number code we were going to be displaying,
which was 127). As I began to build out the program, I also realized
that the number of characters that were to be displayed per line (16),
and the beginning and ending ranges of ASCII number codes/characters to
be displayed (32 to 127) were also invariables.

The biggest learnings for me here were:
- Solve problems on paper first. This allows you to focus on solving the
  problem without get into the implementation details in a particular
  language. For me, this led to less frustration, and as a corollary,
  more satisfaction, during the implementation stage.
- Try to identify variables and invariables during the problem solving
  stage.
- Accept that there will always be a degree of iteration and debugging
  no matter how much work you do in the "solve on paper" stage.